+++ db/migrate/20211029142607_add_postgraduate_loan_repayment_to_payments.rb
@@ -1,0 +1,5 @@
+class AddPostgraduateLoanRepaymentToPayments < ActiveRecord::Migration[6.0]
+  def change
+    add_column :payments, :postgraduate_loan_repayment, :decimal, precision: 7, scale: 2
+  end
+end