+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog]
 - Remove Geckoboard
 - TSLR
   - How much student loan did you repay page content change
+- All policies
+  - Adding postgraduate loan contribution to payroll confirmation
 
 ## [Release 102] - 2021-10-28